refactor: changing project structure
