+++ src/app/app.routes.ts
@@ -1,4 +1,4 @@
 import { Routes } from '@angular/router';
-import { DashboardComponent } from './components/dashboard/dashboard.component';
+import { DashboardComponent } from './pages/dashboard/dashboard.component';
 
 export const routes: Routes = [{ path: '', component: DashboardComponent }];
