+++ src/app/core/components/navbar/navbar.component.scss
